+++ src/graysvr/CItem.cpp
@@ -682,11 +682,11 @@ int CItem::IsWeird() const
 	return( ( ptCont == NULL ) ? 0x2106 : ptCont->IsWeird() );
 }
 
-signed char CItem::GetFixZ( CPointMap pt, unsigned long wBlockFlags )
+signed char CItem::GetFixZ( CPointMap pt, unsigned int wBlockFlags )
 {
 	height_t zHeight = CItemBase::GetItemHeight( GetDispID(), wBlockFlags );
 	CGrayMapBlockState block( wBlockFlags, pt.m_z, pt.m_z + zHeight, pt.m_z + 2, zHeight );
-	g_World.GetFixPoint( pt, block);
+	g_World.GetFixPoint( pt, block );
 	return(block.m_Bottom.m_z);
 }
